requirements.txt
google-cloud-dialogflow==2.13.0
python-dotenv==0.20.0
python-telegram-bot==13.11
vk_api==11.9.7